Added random port selection.
Eventually I need to implement an endpoint on the API that the client can ping
to obtain in available port within the proper range. However, until I can get
around to that this will do. I also added some instructions and a better user
experience to the client since there currently are no notifications.